requirements.txt
docopt==0.6.2
lxml==3.4.0
PyYAML==3.11

# testing
mock==1.0.1
